+++ tests/test_clip_intensity_percentilesd.py
@@ -13,6 +13,7 @@ from __future__ import annotations
 
 import unittest
 
+import torch
 from parameterized import parameterized
 
 from monai.transforms import ClipIntensityPercentilesd
@@ -60,7 +61,7 @@ class TestClipIntensityPercentilesd2D(NumpyImageTestCase2D):
         im = p(self.imt)
         result = soft_clipper({key: im})
         lower, upper = percentile(im, (5, 95))
-        expected = soft_clip(im, sharpness_factor=1.0, minv=lower, maxv=upper, dtype=im.dtype)
+        expected = soft_clip(im, sharpness_factor=1.0, minv=lower, maxv=upper, dtype=torch.float32)
         # the rtol is set to 1e-6 because the logaddexp function used in softplus is not stable accross torch and numpy
         assert_allclose(result[key], p(expected), type_test="tensor", rtol=1e-6, atol=0)
 
@@ -71,7 +72,7 @@ class TestClipIntensityPercentilesd2D(NumpyImageTestCase2D):
         im = p(self.imt)
         result = soft_clipper({key: im})
         upper = percentile(im, 95)
-        expected = soft_clip(im, sharpness_factor=1.0, minv=None, maxv=upper, dtype=im.dtype)
+        expected = soft_clip(im, sharpness_factor=1.0, minv=None, maxv=upper, dtype=torch.float32)
         # the rtol is set to 5e-5 because the logaddexp function used in softplus is not stable accross torch and numpy
         assert_allclose(result[key], p(expected), type_test="tensor", rtol=5e-5, atol=0)
 
@@ -82,7 +83,7 @@ class TestClipIntensityPercentilesd2D(NumpyImageTestCase2D):
         im = p(self.imt)
         result = soft_clipper({key: im})
         lower = percentile(im, 5)
-        expected = soft_clip(im, sharpness_factor=1.0, minv=lower, maxv=None, dtype=im.dtype)
+        expected = soft_clip(im, sharpness_factor=1.0, minv=lower, maxv=None, dtype=torch.float32)
         # the rtol is set to 1e-6 because the logaddexp function used in softplus is not stable accross torch and numpy
         assert_allclose(result[key], p(expected), type_test="tensor", rtol=1e-6, atol=0)
 
@@ -162,7 +163,7 @@ class TestClipIntensityPercentilesd3D(NumpyImageTestCase3D):
         im = p(self.imt)
         result = soft_clipper({key: im})
         lower, upper = percentile(im, (5, 95))
-        expected = soft_clip(im, sharpness_factor=1.0, minv=lower, maxv=upper, dtype=im.dtype)
+        expected = soft_clip(im, sharpness_factor=1.0, minv=lower, maxv=upper, dtype=torch.float32)
         # the rtol is set to 1e-6 because the logaddexp function used in softplus is not stable accross torch and numpy
         assert_allclose(result[key], p(expected), type_test="tensor", rtol=1e-6, atol=0)
 
@@ -173,7 +174,7 @@ class TestClipIntensityPercentilesd3D(NumpyImageTestCase3D):
         im = p(self.imt)
         result = soft_clipper({key: im})
         upper = percentile(im, 95)
-        expected = soft_clip(im, sharpness_factor=1.0, minv=None, maxv=upper, dtype=im.dtype)
+        expected = soft_clip(im, sharpness_factor=1.0, minv=None, maxv=upper, dtype=torch.float32)
         # the rtol is set to 5e-5 because the logaddexp function used in softplus is not stable accross torch and numpy
         assert_allclose(result[key], p(expected), type_test="tensor", rtol=5e-5, atol=0)
 
@@ -184,7 +185,7 @@ class TestClipIntensityPercentilesd3D(NumpyImageTestCase3D):
         im = p(self.imt)
         result = soft_clipper({key: im})
         lower = percentile(im, 5)
-        expected = soft_clip(im, sharpness_factor=1.0, minv=lower, maxv=None, dtype=im.dtype)
+        expected = soft_clip(im, sharpness_factor=1.0, minv=lower, maxv=None, dtype=torch.float32)
         # the rtol is set to 1e-6 because the logaddexp function used in softplus is not stable accross torch and numpy
         assert_allclose(result[key], p(expected), type_test="tensor", rtol=1e-6, atol=0)
